refactor: Separate private modules rules, otlp
